All features implemented now!

+++ src/Processing/FuseImages.java
@@ -17,10 +17,10 @@ public class FuseImages extends ImageProcessor{
 
     private ImageCanvas canvas_1;
     private ImageCanvas canvas_2;
-    private int weight_1;
-    private int weight_2;
+    private double weight_1;
+    private double weight_2;
 
-    public FuseImages(ImageCanvas canvas_1, ImageCanvas canvas_2, int weight_1, int weight_2, PropertyChangeListener plc)
+    public FuseImages(ImageCanvas canvas_1, ImageCanvas canvas_2, double weight_1, double weight_2, PropertyChangeListener plc)
     {
         super(canvas_1);
         this.canvas_1 = canvas_1;
@@ -67,14 +67,12 @@ public class FuseImages extends ImageProcessor{
             for (int y = 0; y < fh; y++) {
                 int rgb_1 = img_1.getRGB(x, y);
                 int rgb_2 = img_2.getRGB(x, y);
-
-                int fr = weight_1 * RGBColor.extractR(rgb_1) +
-                        weight_2 * RGBColor.extractR(rgb_2);
-                int fg = weight_1 * RGBColor.extractG(rgb_1) +
-                        weight_2 * RGBColor.extractG(rgb_2);
-                int fb = weight_1 * RGBColor.extractB(rgb_1) +
-                        weight_2 * RGBColor.extractB(rgb_2);
-
+                int fr = (int)(weight_1 * RGBColor.extractR(rgb_1) +
+                        weight_2 * RGBColor.extractR(rgb_2));
+                int fg = (int)(weight_1 * RGBColor.extractG(rgb_1) +
+                        weight_2 * RGBColor.extractG(rgb_2));
+                int fb = (int)(weight_1 * RGBColor.extractB(rgb_1) +
+                        weight_2 * RGBColor.extractB(rgb_2));
                 fused.setRGB(x, y, RGBColor.combineRGB(fr, fg, fb));
             }
         }
